Refactor app structure into modular screens and services
